Better code

+++ src/dqn/atari_wrappers.py
@@ -122,7 +122,7 @@ def _process_frame84(frame):
 class ProcessFrame84(gym.Wrapper):
     def __init__(self, env=None):
         super(ProcessFrame84, self).__init__(env)
-        self.observation_space = spaces.Box(low=0, high=255, shape=(84, 84, 1))
+        self.observation_space = spaces.Box(low=0, high=255, shape=(84, 84, 1), dtype=np.uint8)
 
     def _step(self, action):
         obs, reward, done, info = self.env.step(action)
